+++ activesupport/CHANGELOG.md
@@ -1,3 +1,9 @@
+*   Remove unused parameter `options = nil` for `#clear` of
+    `ActiveSupport::Cache::Strategy::LocalCache::LocalStore` and
+    `ActiveSupport::Cache::Strategy::LocalCache`.
+
+    *Yosuke Kabuto*
+
 *   Support parsing JSON time in ISO8601 local time strings in
     `ActiveSupport::JSON.decode` when `parse_json_times` is enabled.
     Strings in the format of `YYYY-MM-DD hh:mm:ss` (without a `Z` at